ultimo suspiro

+++ routes/web.php
@@ -14,24 +14,20 @@ use App\Http\Controllers\PagesController;
 //Route::get('/user/{id}', [PagesController::class,'show']);
  
 Route::get('/',[PagesController::class,'fnIndex'] )->name('xIndex');
-Route::post('/',[PagesController::class,'fnRegistrar'] )->name('Estudiante.xRegistrar');
+Route::middleware(['auth:sanctum','verified'])->post('/',[PagesController::class,'fnRegistrar'] )->name('Estudiante.xRegistrar');
 
-Route::get('/detalle/{id}',[PagesController::class,'fnEstDetalle'] )->name('Estudiante.xDetalle');
-Route::get('/lista',[PagesController::class,'fnLista'])->name('xLista');
-Route::get('/galeria/{numero?}', [PagesController::class,'fnGaleria']) -> where('numero','[0-9]+')->name('xGaleria');
+Route::middleware(['auth:sanctum','verified'])->get('/detalle/{id}',[PagesController::class,'fnEstDetalle'] )->name('Estudiante.xDetalle');
+Route::middleware(['auth:sanctum','verified'])->get('/dashboard',[PagesController::class,'fnLista'])->name('dashboard');
+Route::middleware(['auth:sanctum','verified'])->get('/galeria/{numero?}', [PagesController::class,'fnGaleria']) -> where('numero','[0-9]+')->name('xGaleria');
 
 //update
-Route::get('/actualizar/{id}',[PagesController::class,'fnEstActualizar']) ->name('Estudiante.xActualizar');
-Route::put('/actualizar/{id}',[PagesController::class,'fnUpdate'])->name('Estudiante.xUpdate');
+Route::middleware(['auth:sanctum','verified'])->get('/actualizar/{id}',[PagesController::class,'fnEstActualizar']) ->name('Estudiante.xActualizar');
+Route::middleware(['auth:sanctum','verified'])->put('/actualizar/{id}',[PagesController::class,'fnUpdate'])->name('Estudiante.xUpdate');
 //delete
-Route::delete('/eliminar/{id}',[PagesController::class,'fnEliminar'])->name('Estudiante.xEliminar');
+Route::middleware(['auth:sanctum','verified'])->delete('/eliminar/{id}',[PagesController::class,'fnEliminar'])->name('Estudiante.xEliminar');
 
-Route::middleware([
-    'auth:sanctum',
-    config('jetstream.auth_session'),
-'verified'
-])->group(function () {
-    Route::get('/dashboard', function () {
-        return view('dashboard');
-    })->name('dashboard');
-});
+/*
+Route::middleware(['auth:sanctum','verified'])->get('/dashboard', function (){
+    return view('dashboard');
+})->name('dashboard');
+*/
